Move shared tests
